# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/classrooms": {
      "post": {
        "operationId": "create",
        "summary": "Creates and returns a new classroom, given ",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Returns new Classroom with a token"
          }
        },
        "tags": [
          "classrooms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "getClassrooms",
        "summary": "Returns all classrooms tied to a specific user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of Classrooms"
          }
        },
        "tags": [
          "classrooms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/classrooms/{id}": {
      "get": {
        "operationId": "getClassroom",
        "summary": "Returns a classroom by specific ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Classroom details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassroomDto"
                }
              }
            }
          }
        },
        "tags": [
          "classrooms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "deleteClassroom",
        "summary": "Deletes a classroom by specific ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Classroom deleted"
          }
        },
        "tags": [
          "classrooms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/classrooms/{id}/leave": {
      "put": {
        "operationId": "removeStudent",
        "summary": "Removes a user from classroom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "classrooms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/classrooms/{id}/assignments": {
      "get": {
        "operationId": "getAssignments",
        "summary": "Get all assignments to classroom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assignments returned"
          }
        },
        "tags": [
          "classrooms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/classrooms/{id}/regenerate-token": {
      "put": {
        "operationId": "regenerateToken",
        "summary": "Regenerate token by specific classroom ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Classroom updated with new token"
          }
        },
        "tags": [
          "classrooms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/classrooms/{id}/students": {
      "get": {
        "operationId": "getStudents",
        "summary": "Returns all students in a classroom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of Students"
          }
        },
        "tags": [
          "classrooms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/classrooms/{id}/students/{sid}": {
      "delete": {
        "operationId": "removeStudentFromClassroom",
        "summary": "Removes a student from a classroom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "classrooms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/assignments": {
      "post": {
        "operationId": "addAssignment",
        "summary": "Add an assignment to classroom",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssignmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Assignment was added"
          }
        },
        "tags": [
          "assignments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/assignments/{id}": {
      "get": {
        "operationId": "getAssignment",
        "summary": "Returns a assignment by specific ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "assignment details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentDto"
                }
              }
            }
          }
        },
        "tags": [
          "assignments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "deleteAssignment",
        "summary": "Deletes a assignment by specific ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "assignment deleted"
          }
        },
        "tags": [
          "assignments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "update",
        "summary": "Update a assignment by specific ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assignment details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentDetailsDto"
                }
              }
            }
          }
        },
        "tags": [
          "assignments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Bonfire Backend API",
    "description": "Project for CSCC01",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "in": "header",
        "type": "http"
      }
    },
    "schemas": {
...
      "ClassroomDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "token": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "creator_id": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "token",
          "name",
          "created_at",
          "updated_at",
          "creator_id"
        ]
      },
      "CreateAssignmentDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "desc": {
            "type": "string"
          },
          "submit_url": {
            "type": "string"
          },
          "creator_id": {
            "type": "number"
          },
          "classroom_id": {
            "type": "number"
          },
          "due_date": {
            "format": "date-time",
            "type": "string"
          },
          "available_date": {
            "format": "date-time",
            "type": "string"
          },
          "published_date": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "title",
          "desc",
          "submit_url",
          "creator_id",
          "classroom_id",
          "due_date",
          "available_date",
          "published_date"
        ]
      },
      "AssignmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "desc": {
            "type": "string"
          },
          "submit_url": {
            "type": "string"
          },
          "creator_id": {
            "type": "number"
          },
          "classroom_id": {
            "type": "number"
          },
          "due_date": {
            "format": "date-time",
            "type": "string"
          },
          "available_date": {
            "format": "date-time",
            "type": "string"
          },
          "published_date": {
            "format": "date-time",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "title",
          "desc",
          "submit_url",
          "creator_id",
          "classroom_id",
          "due_date",
          "available_date",
          "published_date",
          "created_at",
          "updated_at",
          "deleted"
        ]
      },
      "AssignmentDetailsDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "desc": {
            "type": "string"
          },
          "submit_url": {
            "type": "string"
          },
          "due_date": {
            "format": "date-time",
            "type": "string"
          },
          "available_date": {
            "format": "date-time",
            "type": "string"
          },
          "published_date": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "title",
          "desc",
          "submit_url",
          "due_date",
          "available_date",
          "published_date"
        ]
      }
    }
  }
}